Bump phpstan/phpstan from 1.3.0 to 1.4.5

+++ app/model/Payment/ReadModel/QueryHandlers/RepaymentCandidateListQueryHandler.php
@@ -64,7 +64,7 @@ final class RepaymentCandidateListQueryHandler
     {
         $participantsPayments = array_filter(
             array_column($this->participants->findByCamp($campId), 'repayment', 'personId'),
-            fn (int $amount) => $amount > 0
+            fn (float $amount) => $amount > 0
         );
 
         foreach ($repayments as $repayment) {
